+++ lib/src/di/app_state/separate_run_override.dart
@@ -1,0 +1,24 @@
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+import '../../backend/storage/domain/usecase/separate_run.dart';
+import '../storage.dart';
+import 'current_target_game.dart';
+
+part 'separate_run_override.g.dart';
+
+@riverpod
+class SeparateRunOverride extends _$SeparateRunOverride {
+  @override
+  bool? build() {
+    final repository = ref.watch(sharedPreferenceStorageProvider);
+    final currentGame = ref.watch(targetGameProvider);
+    return initializeSeparateRunOverrideUseCase(repository, currentGame);
+  }
+
+  void setValue(final bool? value) {
+    final repository = ref.read(sharedPreferenceStorageProvider);
+    final currentGame = ref.read(targetGameProvider);
+    setSeparateRunOverrideUseCase(repository, currentGame, value);
+    state = value;
+  }
+}
