+++ lib/src/ui/route/nahida_store/nahida_store.dart
@@ -8,7 +8,7 @@ import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 
-import '../../../backend/akasha/domain/entity/akasha.dart';
+import '../../../backend/akasha/domain/entity/nahida_element.dart';
 import '../../../backend/fs_interface/data/helper/path_op_string.dart';
 import '../../../backend/structure/entity/mod_category.dart';
 import '../../../di/fs_watcher.dart';
