+++ lib/src/ui/route/welcome.dart
@@ -3,7 +3,7 @@ import 'package:flutter/gestures.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../backend/app_version/domain/github.dart';
-import '../../di/app_state.dart';
+import '../../di/app_state/current_target_game.dart';
 import '../util/open_url.dart';
 
 /// This route welcomes you!
